Orgainsed methods into a 'Common' folder and moved submethods into parents method files
