+++ DFC.Common.SharedContent.Pkg.Netcore/Model/Common/Breadcrumb.cs
@@ -1,7 +1,14 @@
 ﻿using System.Text.Json.Serialization;
+using DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems;
 
-namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems
+namespace DFC.Common.SharedContent.Pkg.Netcore.Model.Common
 {
+    public class Breadcrumb
+    {
+        [JsonPropertyName("termContentItems")]
+        public List<TermContentItem>? TermContentItems { get; set; }
+    }
+
     public class TermContentItem
     {
         [JsonPropertyName("displayText")]
