+++ DFC.Common.SharedContent.Pkg.Netcore/Model/Common/Flow.cs
@@ -1,6 +1,7 @@
 ﻿using System.Text.Json.Serialization;
+using DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems;
 
-namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems
+namespace DFC.Common.SharedContent.Pkg.Netcore.Model.Common
 {
     public class Flow
     {
