+++ DFC.Common.SharedContent.Pkg.Netcore/Model/Common/Herobanner.cs
@@ -1,8 +1,8 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems
+namespace DFC.Common.SharedContent.Pkg.Netcore.Model.Common
 {
-    public class TriageToolSummary
+    public class Herobanner
     {
         [JsonPropertyName("html")]
         public string? Html { get; set; }
