+++ DFC.Common.SharedContent.Pkg.Netcore/Model/Common/JobProfileCategory.cs
@@ -1,11 +1,12 @@
-﻿using Newtonsoft.Json;
+﻿using DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems
+namespace DFC.Common.SharedContent.Pkg.Netcore.Model.Common
 {
     public class JobProfileCategory
     {
