+++ DFC.Common.SharedContent.Pkg.Netcore/Model/Common/TriageToolSummary.cs
@@ -1,8 +1,8 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems
+namespace DFC.Common.SharedContent.Pkg.Netcore.Model.Common
 {
-    public class Herobanner
+    public class TriageToolSummary
     {
         [JsonPropertyName("html")]
         public string? Html { get; set; }
