--- DFC.Common.SharedContent.Pkg.Netcore/Model/ContentItems/Breadcrumb.cs
@@ -1,10 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems
-{
-    public class Breadcrumb
-    {
-        [JsonPropertyName("termContentItems")]
-        public List<TermContentItem>? TermContentItems { get; set; }
-    }
-}
