--- DFC.Common.SharedContent.Pkg.Netcore/Model/ContentItems/Dysac/Trait.cs
@@ -1,16 +1,0 @@
-﻿using DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems.Dysac.PersonalityTrait;
-using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems.Dysac
-{
-    public class Trait
-    {
-        [JsonProperty("contentItems")]
-        public List<PersonalityTrait.PersonalityTrait> ContentItems { get; set; }
-    }
-}
