--- DFC.Common.SharedContent.Pkg.Netcore/Model/ContentItems/FormElement.cs
@@ -1,10 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems
-{
-    public class FormElement
-    {
-        [JsonProperty("id")]
-        public string? Id { get; set; }
-    }
-}
