+++ DFC.Common.SharedContent.Pkg.Netcore/Model/ContentItems/JobProfile.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using DFC.Common.SharedContent.Pkg.Netcore.Model.Common;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -73,5 +74,23 @@ namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems
 
         [JsonProperty("relatedSkill")]
         public string? RelatedSkillDesc { get; set; }
+    }
+
+    public class WorkingPattern
+    {
+        [JsonProperty("contentItems")]
+        public List<ContentItem> ContentItems { get; set; }
+    }
+
+    public class WorkingHoursDetails
+    {
+        [JsonProperty("contentItems")]
+        public List<ContentItem> ContentItems { get; set; }
+    }
+
+    public class WorkingPatternDetails
+    {
+        [JsonProperty("contentItems")]
+        public List<ContentItem> ContentItems { get; set; }
     }
 }
