--- DFC.Common.SharedContent.Pkg.Netcore/Model/ContentItems/WorkingHoursDetails.cs
@@ -1,10 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems
-{
-    public class WorkingHoursDetails
-    {
-        [JsonProperty("contentItems")]
-        public List<ContentItem> ContentItems { get; set; }
-    }
-}
