--- DFC.Common.SharedContent.Pkg.Netcore/Model/ContentItems/WorkingPattern.cs
@@ -1,10 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems
-{
-    public class WorkingPattern
-    {
-        [JsonProperty("contentItems")]
-        public List<ContentItem> ContentItems { get; set; }
-    }
-}
