--- DFC.Common.SharedContent.Pkg.Netcore/Model/ContentItems/WorkingPatternDetails.cs
@@ -1,10 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems
-{
-    public class WorkingPatternDetails
-    {
-        [JsonProperty("contentItems")]
-        public List<ContentItem> ContentItems { get; set; }
-    }
-}
