+++ DFC.Common.SharedContent.Pkg.Netcore/Model/Response/TriageToolFilterResponse.cs
@@ -1,4 +1,4 @@
-﻿using DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems;
+﻿using DFC.Common.SharedContent.Pkg.Netcore.Model.Common;
 using DFC.Common.SharedContent.Pkg.Netcore.Model.ContentItems.PageBanner;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
